Added enemy: Wizard
Currently the wizard acts like the skeleton but it can shoot. I want to change this behavior so it keeps some range. It also has infinite shooting range and that need to be fixed. Additionally it needs a bullet sprite, depending on what we do with the sprites I may create this.